Fork off the insights endpoints and tests.
