+++ server/integration_tests/insights_test.py
@@ -1,0 +1,120 @@
+# Copyright 2023 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import json
+import os
+
+import requests
+
+from shared.lib.test_server import NLWebServerTestCase
+
+_dir = os.path.dirname(os.path.abspath(__file__))
+
+_TEST_MODE = os.environ['TEST_MODE']
+
+_TEST_DATA = 'test_data'
+
+
+class IntegrationTest(NLWebServerTestCase):
+
+  def run_fulfillment(self, test_dir, req_json, failure=''):
+    resp = requests.post(self.get_server_url() + '/api/insights/fulfill',
+                         json=req_json).json()
+    self.handle_response(json.dumps(req_json), resp, test_dir, '', failure)
+
+  def run_detection(self, test_dir, query, failure=''):
+    resp = requests.post(self.get_server_url() +
+                         f'/api/insights/detect?q={query}',
+                         json={}).json()
+    self.handle_response(query, resp, test_dir, '', failure)
+
+  def handle_response(self,
+                      query,
+                      resp,
+                      test_dir,
+                      test_name,
+                      failure,
+                      check_place_detection=False,
+                      detector=None):
+    dbg = resp['debug']
+    resp['debug'] = {}
+    resp['context'] = {}
+    json_file = os.path.join(_dir, _TEST_DATA, test_dir, test_name,
+                             'chart_config.json')
+    if _TEST_MODE == 'write':
+      json_dir = os.path.dirname(json_file)
+      if not os.path.isdir(json_dir):
+        os.makedirs(json_dir)
+      with open(json_file, 'w') as infile:
+        infile.write(json.dumps(resp, indent=2))
+
+      if check_place_detection:
+        dbg_file = os.path.join(json_dir, 'debug_info.json')
+        with open(dbg_file, 'w') as infile:
+          dbg_to_write = {
+              "places_detected": dbg["places_detected"],
+              "places_resolved": dbg["places_resolved"],
+              "main_place_dcid": dbg["main_place_dcid"],
+              "main_place_name": dbg["main_place_name"]
+          }
+          infile.write(json.dumps(dbg_to_write, indent=2))
+    else:
+      if failure:
+        self.assertTrue(failure in resp["failure"]), resp["failure"]
+        self.assertTrue(not resp["config"])
+        return
+
+      if detector:
+        self.assertTrue(dbg.get('detection_type').startswith(detector)), \
+          f'Query {query} failed!'
+      if not check_place_detection:
+        with open(json_file, 'r') as infile:
+          expected = json.load(infile)
+          expected['debug'] = {}
+          expected['context'] = {}
+          a, b = (
+              json.dumps(resp, sort_keys=True, indent=2),
+              json.dumps(expected, sort_keys=True, indent=2),
+          )
+          self.maxDiff = None
+          self.assertEqual(a, b)
+      else:
+        # Look in the debugInfo file to match places detected.
+        dbg_file = os.path.join(_dir, _TEST_DATA, test_dir, test_name,
+                                'debug_info.json')
+        with open(dbg_file, 'r') as infile:
+          expected = json.load(infile)
+          self.assertEqual(dbg["places_detected"], expected["places_detected"])
+          self.assertEqual(dbg["places_resolved"], expected["places_resolved"])
+          self.assertEqual(dbg["main_place_dcid"], expected["main_place_dcid"])
+          self.assertEqual(dbg["main_place_name"], expected["main_place_name"])
+
+  def test_detection_basic(self):
+    self.run_detection('detection_api_basic', 'Commute in California')
+
+  def test_detection_childtype(self):
+    self.run_detection('detection_api_childtype',
+                       'Commute in counties of California')
+
+  def test_fulfillment_basic(self):
+    req = {'entities': ['geoId/06'], 'variables': ['dc/topic/WorkCommute']}
+    self.run_fulfillment('fulfillment_api_basic', req)
+
+  def test_fulfillment_childtype(self):
+    req = {
+        'entities': ['geoId/06'],
+        'variables': ['dc/topic/WorkCommute'],
+        'childEntityType': 'County'
+    }
+    self.run_fulfillment('fulfillment_api_childtype', req)
